Creating model class to hold the 'data'
